Fixes Gentics Support Ticket #43496 - splitting, joining, splitting causes text to be lost
Let there be a paragraph, or another splittable element that contains
text. Let the element be split in the middle of the contained text - the
result will be two elements, each of which containing part of the
splitted text.

Let the two elements be joined again with backspace - on IE7 the result
will be one element that contains two text nodes.

On IE7, if the splittable element with two text nodes was split again,
the second text node was lost.