Fixes disappearing repository links (e.g. internal links)
The link plugin will automatically remove a link if no text was entered
in the ui-attributefield.

Links that are entered literally in the ui-attributefield worked
flawlessly. Links arrvied at through a link repository, however, were
removed as soon as one left the ui-attributefield (blur), since there
was no check for whether a resource was entered.