!!! init() now requires postProcessFn as second parameter; init MUST BE idempotent now!
When working with nested aloha blocks, it can happen (for various reasons)
that this.$element of a block *looses* connection to the DOM element which
is currently rendered.
Each time we detect this, we call _connectThisBlockToDomElement(), which sets
up the connection again and also calls init() again.

Thus, init() must be idempotent AND call postProcessFn for working correctly.