* Use jQuery event delegation to bind link events to newly added links objects in editable area.
* Use `inserthtml` command to insert a new link to a collapsed range.

* Use  `createLink` command create a new link from a selection.

* Update range selection after a link is added to the editable area.

* Use `unlink` command to remove a link.

* Actually, `unlink` command only resets the `href` tag of anchor(`a`)
tag. Rest of the attributes are left intact with the tag.
However, user's intention of using removeLink is to remove the `a` tag
visually.
To achieve this behaviour, we do a cleanup which converts the `a` tag to
a `span`,
while keeping the original attributes intact (class values set by the
Link plugin are removed).'