Fixes a bug that caused only a single root node to be displayed
The if statement was supposed to be part of a fix for duplicate browser entries in commit
508d7a6c0278210d56aa2322d0bc09d75223e7de

The if statement caused only a single root node to be displayed with the
GCN repository. I was unable to reproduce the bug with duplicate browser
entries after removing the if, so maybe it was optional?