We have found that the following comment
//@{ 

is actually parsed by some versions of IE9 and IE8 in our pages, resulting in the following error:  

SCRIPT1009: Expected '}' 
testutils.js, line XXX character 2

It is as though IE parses the "{" character of the comment as the beginning of a script block, and then complains that it isn't closed.

Removing this comment fixes the bug for us.  Please consider eliminating the "{" character from future comments.