Implementing work-around for IE7 bug in `isCollapsedLineBreak()'.
The IE7 behaviour of getAttribute() is buggy in that it differs from the
documentation.  According to the DOM Level 2 documentation at:
http://msdn.microsoft.com/en-us/library/ie/ms536429(v=vs.85).aspx, the return
value of getAttribute() should be a string, interger, or boolean.  But when one
queries for node.getAttribute('style'), what is actually returned is not a
string, as one might expect, but an object.  The returned object is in fact the
style *object*.  I.e., node.style === node.getAttribute('style') .