Fixes autocomplete loading errors
The jquery-plugin requirejs plugin that loads jquery
plugins via an ajax request somtimes causes errors.
I thought switching it to async:false in the jquery-plugin.js
fixes this error, because it seemed that way during the
wikidocs integration, but it was still causing errors
in this branch.

Also, the jquery-plugin requirejs plugin keeps causing
cors errors (cross origin request s?) during loading, but
only somtimes, so the jquery-plugin should be phased out
and only define wrappers should be used in the future.