Ports the link-plugin to use the ui-plugin
With this the link-plugin ui is completely implemented with the ui-plugin.

Special care has been taken that all original functionality was preserved.

The following code didn't do anything as far as I could tell, so I just
replaced it with a single call to Toolbar.activateTabOfButton()

// remember the current tab selected by the user
var currentTab = FloatingMenu.userActivatedTab;

// switch to the href tab (so that we make sure that the href field gets created)
FloatingMenu.activateTabOfButton( 'href' );
if ( currentTab ) {
    // switch back to the original tab
    FloatingMenu.userActivatedTab = currentTab;
}