Fixes loading errors for IE8
There was a loading error in ui/toolbar - the define callback was
invoked with all dependencies except jQuery being undefined.

A hunch caused me to remove the order! plugin from the ui-plugin
dependency on toolbar. That made the loading errors go away.

It is unknown why this caused problems in IE8. It worked in Chrome.
The following may give some insight:

https://github.com/jrburke/requirejs/issues/121
So, the the order plugin has been removed and following the lead of Tim Branyen and Dave Geddes, of use and wrap respectively, requirejs 2.0 integrates that kind of dependency tree specification directly in requirejs.
https://github.com/jrburke/requirejs/wiki/Upgrading-to-RequireJS-2.0

Since the order plugin is apparently not a good idea anyway (removed in 2.0)
I removed all uses of the order plugin.

The css dependencies for jqueryui and ui require to be in-order due to
classes with conflicting styles but same priority. I moved them to the
global aloha.css include. That is probably a better idea anyway than
to use another requijs plugin which may or may not work reliably across
all browsers.