Fixes block draghandles are sometimes missing
Occured with blocks nested inside an editable for
which dragndrop was activated, which was again nested
inside a block for which dragndrop was no activated.

Actually, in the case above, the outer block had
dragndrop disabled due to a bug, which is also fixed
with this commit.

This only occurred sometimes due to a timing issue -
when first the editable was initialized and then the
outer block, the error would occur. The order in which
blocks are initialized is not defined however, and will
not always be the same, probably due to the generous use
of setTimeout and spaghetti in this implementation.