Adds to Aloha Block initialization to publish on "aloha.blocks.initialized" channel when a block is fully initialized.
Aloha Blocks had no facility to notify when once a block has been fully
initialized.  The initialization of an Aloha Block invovles
instanciating a Block object in memory, associating this object with a
corresponding DOM element, decorating the DOM elment with scaffolding
UI, registering events for the new block instance, and registering the
instance with the BlockManager.

With this patch, a message will be published on
"aloha.blocks.initialized" when the above initialization process is
completed for each new block instance.