Standardize on 'repository item'.
Until before this commit, 'repository item' and 'resource item' would be used interchangeably. Even to the point that code examples are in fact broken because the parameter is called 'repositoryItem' yet when it is used, it is called 'resourceItem'. We can still rename it later on, but this at least makes the naming consistent, preventing confusion for users/readers.