Handle ui-break in dialog better
An empty string now indicates that we require a component to be placed
on a new line. When such an empty string is seen, an html break is added.
This makes dialog layout slightly more flexible. Also add hook for dialog close
handler, which needs to change focus away from our current scope (somehow).