Register components other than buttons, react to scope changes
Scope changes are deprecated, but adding it is so easy, might as well
support it for now.

If a plugin registers something for which we can find a component in the
toolbar, place that component there. This allows the use of "tabs" of
components registered by plugins (eg: Image plugin), and renders it as
a dialog. A future version will instead delegate this to a bubble with a button
that brings up the dialog.