Implements asynchronous plugin initialization
Aloha has assumed that a plugin is finished initializing when it's
init() method returns--not taking into account the possibility of
asynchronous processes in plugin initialization.  This oversight allowed
for the "aloha-ready" event to be fired before plugins that required
asynchronous initialization were fully initialized.

This patch allows plugins to return a $.Deferred() object from their
init method in order to allow for their initialization to be handled
asynchronously by the PluginManager, and thereby defer the "aloha-ready"
event until the plugin has fully finished its initialized process.