Overhaul Aloha's initialization process
Aloha has had a initialization process which was not very exensible.
Initialization is now made to more robust and flexible.

Users expect that when their Aloha.ready() handlers are fired, no plugin
is going to manipulate the editables for the purpose of their
initialization process.

In order to establish this contract, plugins need to he able to
manipulate fully alofied editables before the "aloha-ready" event,
during the course of their initialization.

It is for this reason that a new event is added: "aloha-plugins-loaded."
From now on "aloha-plugins-loaded" is the threshold at which editable
can be alofied.  This means that plugins can manipulate the editables
during the course of their initialization, _before_ control is passed to
the user at "aloha-ready."