clean up on all hide events
This ensures that empty math wrappers are always cleaned up, whether you
use the buttons or just click outside.

Also, ensure tooltips are only enabled/disabled if they exist. This is a
workaround for another bug that multiplies the popovers on items and causes
the event handler to be invoked in multiples of two.