fix multiple header adding bug
This occurs because the code that determines what is available only runs
on context switch, so you can add as many headers as you like as long as you
don't context switch (move the cursor). Disabling the menu item also disables
the preview functions, so we clean up after ourselves here.