disable click and preview for disabled items
Before this fix, clicking a disabled item would call its click handler,
which means every plugin needs to deal with the possibility of being called
even if the functions it provides should not be available. Also disable
previewing, no point in showing what would happen if it shouldn't happen,
not to mention, plugins then have to deal with previewing an action that
might not even be possible. Much better to hide this at a higher level.