Fixes de-pressing a format button doesn't work with state-override
De-pressing a format button removed the format from the whole formatted
range, which is different from the behaviour when a format button is
pressed, which only does an extend to word, not an extend to the whole
range.

Unformatting the whole formatted range is incompatible with the
state-override functionality, which requires de-pressing a format button
not to perform any action when between words or at the end of a
sentence.

This fix makes pressing and de-pressing the format button behave the
same: both try an extend to word first, and that doesn't extend the
range, set the state-override.

MS Word behaves similarly.