Link plugin: click once to open link Tab. 'self' as default target
When typing the URL and then pressing ENTER, the link is created but the
user must click two times for the link properties Tab to open. The
changes in this fix allow to open the link properties by clicking just
once.

The default target was set to 'framename' instead of 'self'.