table-plugin: setting a row as table header does remove the row selection.
When setting a row as table header the cell selection was not removed. Setting a row as table header removes the selected row and replaces it by a new one. The problem was that the unselect function was called for the removed row instead of for the newly inserted row. Updating the selection to the new row fixes the issue.