IE default drag and drop behavior is stopped when dragging and dropping block elements.
In some cases the default IE drag and drop behavior was not stopped for
blocks, so dragging and dropping blocks messed up the content. The root
of the problem was a Javascript Error (rangy-core) when dragging the
element. This error was thrown due to a bug in the IE selection. With
this fix the IE default drag and drop behavior is stopped and the
Javascript Error is catch, so the excution of the program can continue.