New check for block-level-elements
Performing unformatting on a selection adjacent to a terminal whitespace sometimes caused a <br> to be erroneously inserted.
For example: inside <sup>example <sup>, if you select "example" without the succeeding space and remove the upper case, then a <br> tag will be added.
The fix checks, if the operation is within a block-level-element and does not add the <br> if it's not a block-level-element.