dom.js: Fix insertIntoDOM() problem, which can cause a new object not to get inserted into the DOM.
The selected range is sometimes incorrect (either a bug in Aloha or IE - so far the problem has only occurred in some versions of IE9). Sometimes the selection offset is greater than the length of the contents of the container, where the object is to be inserted. In that case adjust the offset to the length of the container contents.

SUP-1079