WordContentHandler: Fix span removal in strange lists.
When some paragraphs are formatted as ListParagraph in Word, but have no bullets, numbers, letters, etc., these paragraphs might not contain the span element that is normally used to detect the list type and which is removed after detection (this causes an entire paragraph to be removed in such a case). Only remove this span if it either has an "mso-list: Ignore" style or if it is nested within a <!--[if !supportLists]--> comment.

SUP-1286