Fix (forard)delete command for whitespace at the beginning of block elements (#1604)
* Fix (forard)delete command for whitespace at the beginning of block elements

While adding a test for this case I noticed the the test suit for the "delete"
command has not been executed because of an error in the corresponding test HTML
file. When renabling the tests several tests failed.

I fixed the ones where I could figure out what should happen and removed the
ones that failed despite Aloha-Editor behaving normaly in this case, and the
ones I could not figure out what should happen.